services added

+++ src/main/java/com/example/online_school_is/repos/CoursesRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface CoursesRepository extends JpaRepository<Courses, Integer> {
+public interface CoursesRepository extends JpaRepository<Courses, Long> {
 
     @Query("select p from users p where concat(p.id, '', p.course_name, '', p.description, '', " +
             "p.teacher_id) like %?1%")
